splitting out cordova.platform into smaller modules. factored a few small helper functions that will be used in multiple modules into the util module.
